[MIL-1] refactor: move utils but keep static js utils in public
