+++ src/utils/emoji_generator.js
