+++ tests/utils/language_colors_generator.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-undef */
-const { getColorsForLanguages } = require('../../src/public/js/language_colors_generator');
+const { getColorsForLanguages } = require('../../src/utils/language_colors_generator');
 const expect = require('expect.js');
 
 describe('emojiGenerator', () => {
